Partial attack
Some of the attack functionality added
