+++ Vermilion Game/Assets/EnemyFriction.physicsMaterial2D
@@ -8,4 +8,4 @@ PhysicsMaterial2D:
   m_PrefabAsset: {fileID: 0}
   m_Name: EnemyFriction
   friction: 0
-  bounciness: 0.01
+  bounciness: 0.05
